+++ metadata.rb
@@ -25,7 +25,7 @@ license 'Apache 2.0'
 description 'Installs and configures a mail server using Postfix, Dovecot, '\
             'PostfixAdmin and SpamAssassin, including Amazon SES support.'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version '3.0.0' # WiP
+version '3.0.0'
 
 if respond_to?(:source_url)
   source_url "https://github.com/zuazo/#{name}-cookbook"